+++ NEWS.md
@@ -1,5 +1,7 @@
 # bsplus 0.0.9.9003
 
+* fiddles with label format for shiny inputs (pulls help-links to the right)
+
 * version bump as new repository-version needed
 
 # bsplus 0.0.9.9002